% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mixture.R
\name{gmm_EM}
\alias{gmm_EM}
\title{EM algorithm for Gaussian Mixture Models}
\usage{
gmm_EM(X, M, epsilon, n.iter, tol)
}
\arguments{
\item{X}{The input data matrix.}

\item{M}{Maximum number of mixture components.}

\item{epsilon}{This parameter is used in penalty function, default 1e-6.}

\item{n.iter}{This parameter is the number of iteration for EM algorithm, default 200.}

\item{tol}{This parameter is the tolerance for likelihood function, default 1e-5.}
}
\description{
This function allows you to do Model Selection for Gaussian Mixture Models.
}
\examples{
library(MASS)
M = 10
n <- 900
S1 <- matrix(c(0.65,0.7794,0.7794,1.55),nrow=2,byrow=TRUE)
mu1 <- c(-1,1)
S2 <- matrix(c(0.65,-0.7794,-0.7794,1.55),nrow=2,byrow=TRUE)
mu2 <- c(1,1)
S3 <- matrix(c(2,0,0,0.2),nrow=2,byrow=TRUE)
mu3 <- c(0,-sqrt(2))
mu <- rbind(mu1,mu2,mu3)
S <- list(S1,S2,S3)
prop <- c(1/3,1/3,1/3)
val1 <- mvrnorm(n*prop[1],mu=mu1,Sigma=S1)
val2 <- mvrnorm(n*prop[2],mu=mu2,Sigma=S2)
val3 <- mvrnorm(n*prop[3],mu=mu3,Sigma=S3)
allval <- rbind(val1,val2,val3)
X <- allval[sample(n,n),]
lambda = sqrt(log(n)/n)/12
res = gmm_EM(X, M, epsilon = 1e-6, n.iter=1000, tol=1e-5)
k.hat = res$KA
prop.hat = res$PI
mu.hat = res$MEAN
Sigma.hat = res$VAR
}
\keyword{Gaussian}
\keyword{mixture}
