% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mixture.R
\name{gmm_kmeans}
\alias{gmm_kmeans}
\title{K-means function for Gaussian Mixture Models}
\usage{
gmm_kmeans(X, K)
}
\arguments{
\item{X}{The input data matrix.}

\item{K}{The initial number of mixing components.}
}
\description{
This function allows you to use k-means clustering method to give the initial values for EM algorithm.
}
\keyword{K-means}
\keyword{clustering}
