% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mixture.R
\name{penalty}
\alias{penalty}
\title{Penalty function for Gaussian Mixture Models}
\usage{
penalty(X, prop, epsilon)
}
\arguments{
\item{X}{The input data matrix.}

\item{prop}{The mixing probability.}

\item{epsilon}{This parameter is used in penalty function, default 1e-6.}
}
\description{
This function allows you to calculate the penalty function for Gaussian mixture model.
}
\keyword{Penalty}
\keyword{function}
